+++ account-kit/java/signer-sdk/src/main/java/com/alchemy/aa/Main.java
@@ -1,0 +1,12 @@
+package com.alchemy.aa;
+
+public class Main {
+
+  public static void main(String[] args) {
+    System.out.printf("Hello and welcome!");
+
+    for (int i = 1; i <= 5; i++) {
+      System.out.println("i = " + i);
+    }
+  }
+}