+++ account-kit/react/src/hooks/useDropAndReplaceUserOperation.ts
@@ -7,7 +7,7 @@ import type {
 } from "@aa-sdk/core";
 import type { SupportedAccounts } from "@account-kit/core";
 import { useMutation, type UseMutateFunction } from "@tanstack/react-query";
-import { useAlchemyAccountContext } from "../context.js";
+import { useAlchemyAccountContext } from "./useAlchemyAccountContext.js";
 import { ClientUndefinedHookError } from "../errors.js";
 import type { BaseHookMutationArgs } from "../types.js";
 import type { UseSmartAccountClientResult } from "./useSmartAccountClient.js";
@@ -60,9 +60,8 @@ export type UseDropAndReplaceUserOperationResult<
  * } from "@account-kit/react";
  *
  * export function ComponentWithDropAndReplaceUO() {
- *   const { client } = useSmartAccountClient({
- *     type: "MultiOwnerModularAccount",
- *   });
+ *   const { client } = useSmartAccountClient({});
+ *
  *   const { sendUserOperationAsync, isSendingUserOperation } =
  *     useSendUserOperation({
  *       client,