+++ examples/ui-demo/src/app/config.tsx
@@ -1,4 +1,5 @@
 import { AuthCardHeader } from "@/components/shared/AuthCardHeader";
+import { odyssey, splitOdysseyTransport } from "@/hooks/7702/transportSetup";
 import { alchemy, arbitrumSepolia } from "@account-kit/infra";
 import { cookieStorage, createConfig } from "@account-kit/react";
 import { AccountKitTheme } from "@account-kit/react/tailwind";
@@ -39,8 +40,11 @@ export type Config = {
         }
       | undefined;
   };
+  accountMode: AccountMode;
   supportUrl?: string;
 };
+
+export type AccountMode = "default" | "7702";
 
 export const DEFAULT_CONFIG: Config = {
   auth: {
@@ -70,6 +74,7 @@ export const DEFAULT_CONFIG: Config = {
     logoLight: undefined,
     logoDark: undefined,
   },
+  accountMode: "default",
 };
 
 export const queryClient = new QueryClient();
@@ -79,8 +84,19 @@ export const alchemyConfig = () =>
     {
       transport: alchemy({ rpcUrl: "/api/rpc" }),
       chain: arbitrumSepolia,
+      chains: [
+        {
+          chain: arbitrumSepolia,
+          transport: alchemy({ rpcUrl: "/api/rpc" }),
+          policyId: process.env.NEXT_PUBLIC_PAYMASTER_POLICY_ID,
+        },
+        {
+          chain: odyssey,
+          transport: splitOdysseyTransport,
+          policyId: process.env.NEXT_PUBLIC_PAYMASTER_POLICY_ID,
+        },
+      ],
       ssr: true,
-      policyId: process.env.NEXT_PUBLIC_PAYMASTER_POLICY_ID,
       connectors: [
         walletConnect({ projectId: "30e7ffaff99063e68cc9870c105d905b" }),
       ],